Различные дополнения (PCA, перцептрон и проч.)

+++ include/distance.h
@@ -23,7 +23,7 @@
 \return расстояние между векторами.
 
 */
-double fmll_distance_square_euclid(const double * vec_1, const double * vec_2, uint8_t dim);
+double fmll_distance_square_euclid(const double * vec_1, const double * vec_2, unsigned dim);
 
 /*!
 
@@ -36,7 +36,7 @@ double fmll_distance_square_euclid(const double * vec_1, const double * vec_2, u
 \return расстояние между векторами.
 
 */
-double fmll_distance_euclid(const double * vec_1, const double * vec_2, uint8_t dim);
+double fmll_distance_euclid(const double * vec_1, const double * vec_2, unsigned dim);
 
 /*!
 
@@ -49,7 +49,7 @@ double fmll_distance_euclid(const double * vec_1, const double * vec_2, uint8_t 
 \return расстояние между векторами.
 
 */
-double fmll_distance_scalar(const double * vec_1, const double * vec_2, uint8_t dim);
+double fmll_distance_scalar(const double * vec_1, const double * vec_2, unsigned dim);
 
 /*!
 
@@ -62,7 +62,7 @@ double fmll_distance_scalar(const double * vec_1, const double * vec_2, uint8_t 
 \return расстояние между векторами.
 
 */
-double fmll_distance_manhattan(const double * vec_1, const double * vec_2, uint8_t dim);
+double fmll_distance_manhattan(const double * vec_1, const double * vec_2, unsigned dim);
 
 /*!
 
@@ -75,7 +75,7 @@ double fmll_distance_manhattan(const double * vec_1, const double * vec_2, uint8
 \return расстояние между векторами.
 
 */
-double fmll_distance_chebyshev(const double * vec_1, const double * vec_2, uint8_t dim);
+double fmll_distance_chebyshev(const double * vec_1, const double * vec_2, unsigned dim);
 
 #endif
